REVIEW: porting to FreeBSD 5.0

+++ Source/Sync/CxProcess.cpp
@@ -30,7 +30,7 @@ ULONG
 CxProcess::ulGetCurrId() {
     /*DEBUG*/// n/a
 
-    ULONG ulRes = - 1;
+    ULONG ulRes = (ULONG)- 1;
 
 #if defined(xOS_WIN)
     ulRes = ::GetCurrentProcessId();
